requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2

sqlalchemy
psycopg2-binary

pillow
numpy
faiss-cpu
opencv-python-headless

insightface
onnxruntime
deepface

python-dotenv
pydantic
pydantic-settings

fastapi
fastapi-cors
uvicorn[standard]
sqlalchemy
jinja2
python-multipart
requests

azure-storage-blob